CLI-64: Refactor services logic; polish service config handling
