+++ lib/CLIManager.js
@@ -51,7 +51,6 @@ const ServicesController = require('./service/ServicesController');
 const EnvironmentsController = require('./environment/EnvironmentsController');
 const CollectionsController = require('./collection/CollectionsController');
 
-const FlexService = require('./flex/FlexService');
 const ServicesService = require('./service/ServicesService');
 const OrganizationsService = require('./organization/OrganizationsService');
 const ApplicationsService = require('./application/ApplicationsService');
@@ -87,10 +86,10 @@ class CLIManager {
   _registerControllers() {
     this._controllers = {};
     this._controllers.init = new InitController({ cliManager: this });
-    const flexService = new FlexService(this);
     const applicationsService = new ApplicationsService(this);
     const organizationsService = new OrganizationsService(this);
-    this._controllers[Namespace.FLEX] = new FlexController({ cliManager: this, flexService, applicationsService, organizationsService });
+    const servicesService = new ServicesService(this);
+    this._controllers[Namespace.FLEX] = new FlexController({ cliManager: this, applicationsService, organizationsService, servicesService });
     this._controllers[Namespace.PROFILE] = new ProfilesController({ cliManager: this });
     this._controllers[Namespace.ORG] = new OrganizationsController({ cliManager: this, organizationsService });
     this._controllers[Namespace.APP] = new ApplicationsController({ cliManager: this, applicationsService });
@@ -107,8 +106,8 @@ class CLIManager {
       collectionsService,
       businessLogicService
     });
-    const servicesService = new ServicesService(this);
-    const serviceFileProcessor = new ServiceFileProcessor({ cliManager: this, servicesService, flexService });
+
+    const serviceFileProcessor = new ServiceFileProcessor({ cliManager: this, servicesService });
 
     this._configFileProcessor = new ConfigFileProcessor({ envFileProcessor, serviceFileProcessor });
   }
