+++ lib/CommandResult.js
@@ -18,7 +18,7 @@ const consoleTable = require('console.table');
 const { EOL } = require('os');
 
 const { OperationMessage, OperationType, OutputFormat } = require('./Constants');
-const { isEmpty, isNullOrUndefined } = require('./Utils');
+const { isEmpty, isNullOrUndefined, isStringWhitespace } = require('./Utils');
 
 /**
  * Defines the result from the successful execution of a command.
@@ -30,7 +30,9 @@ class CommandResult {
   constructor() {
     this.rawData = null;
     this.tableData = null;
-    this.humanReadableFormat = {};
+    this.humanReadableFormat = {
+      userFriendlyMsg: ''
+    };
   }
 
   /**
@@ -121,6 +123,16 @@ class CommandResult {
   }
 
   /**
+   * Appends an additional message to the user friendly message.
+   * @param {String} msg
+   * @param {String} [separator]
+   */
+  setAdditionalMsg(msg, separator = EOL) {
+    this.humanReadableFormat.userFriendlyMsg = `${this.humanReadableFormat.userFriendlyMsg}${separator}${msg}`;
+    return this;
+  }
+
+  /**
    * Sets a custom user-friendly message.
    * @param {String} msg
    * @returns {CommandResult}
@@ -135,7 +147,7 @@ class CommandResult {
    * @returns {boolean}
    */
   _humanReadableFormatIsTable() {
-    const isTable = isNullOrUndefined(this.humanReadableFormat.userFriendlyMsg) && (!isNullOrUndefined(this.tableData) || !isNullOrUndefined(this.rawData));
+    const isTable = isStringWhitespace(this.humanReadableFormat.userFriendlyMsg) && (!isNullOrUndefined(this.tableData) || !isNullOrUndefined(this.rawData));
     return isTable;
   }
 }
