+++ lib/ServiceFileProcessor.js
@@ -15,14 +15,13 @@
 
 const async = require('async');
 
-const { FlexProjectMaxSize, OperationType } = require('./Constants');
+const { OperationType } = require('./Constants');
 const { isNullOrUndefined } = require('./Utils');
 
 class ServiceFileProcessor {
   constructor(options) {
     this.cliManager = options.cliManager;
     this.servicesService = options.servicesService;
-    this.flexService = options.flexService;
   }
 
   process(options, done) {
@@ -38,8 +37,6 @@ class ServiceFileProcessor {
   }
 
   _createService(source, options, done) {
-    debugger;
-    // TODO: type should be validated at some point
     const data = {
       name: options.name,
       type: source.type === 'flex-internal' ? 'internal' : 'external',
@@ -64,36 +61,24 @@ class ServiceFileProcessor {
         return done(null, service);
       }
 
-      this.flexService.validateProject(dir, (err, pkgVersion) => {
+      this.servicesService.deployFlexProject(dir, service.id, this.cliManager.config.defaultSchemaVersion, (err, id) => {
         if (err) {
           return done(err);
         }
 
-        this.flexService.deployProject(
-          dir,
-          service.id,
-          pkgVersion,
-          this.cliManager.config.defaultSchemaVersion,
-          this.cliManager.config.flexProjectUploadTimeout,
-          FlexProjectMaxSize,
-          this.cliManager.config.artifacts,
-          (err, deployJobId) => {
-            if (err) {
-              return done(err);
-            }
 
-            console.log(`===Deploy job ID: ${deployJobId}`);
-            done(null, service);
-          }
-        );
+        service.jobId = id;
+        done(null, service);
       });
     });
   }
 
   _updateService(source, options, done) {
     const serviceId = options.serviceId;
+    let updateData;
+    let jobId;
 
-    async.waterfall([
+    async.series([
       (next) => {
         this.servicesService.getById(serviceId, (err, originalService) => {
           if (err) {
@@ -115,18 +100,35 @@ class ServiceFileProcessor {
             data.description = source.description;
           }
 
-          next(null, data);
+          updateData = data;
+
+          next();
         });
       },
-      (data, next) => {
-        this.servicesService.update(data, serviceId, next);
+      (next) => {
+        this.servicesService.update(updateData, serviceId, next);
+      },
+      (next) => {
+        const dir = source.sourcePath;
+        if (!dir) {
+          return setImmediate(next);
+        }
+
+        this.servicesService.deployFlexProject(dir, serviceId, this.cliManager.config.defaultSchemaVersion, (err, id) => {
+          if (err) {
+            return next(err);
+          }
+
+          jobId = id;
+          next();
+        });
       }
-    ], (err, result) => {
+    ], (err) => {
       if (err) {
         return done(err);
       }
 
-      done(null, result);
+      done(null, { id: serviceId, jobId });
     });
   }
 }
