+++ lib/flex/FlexController.js
@@ -15,13 +15,15 @@
 
 const async = require('async');
 const chalk = require('chalk');
+const moment = require('moment');
 
 const BaseController = require('../BaseController');
 const ProjectSetup = require('../ProjectSetup');
-const { AppOptionsName, Command, CommonOptionsNames, Errors, EntityType, DomainTypes, LogLevel, OperationType, OrgOptionsName, PromptTypes, FlexOptionsNames, FlexProjectMaxSize, JobStatus } = require('../Constants');
+const { AppOptionsName, Command, CommonOptionsNames, Errors, EntityType, DomainTypes, InfoMessages, LogErrorMessages, LogLevel,
+  OperationType, OrgOptionsName, PromptMessages, PromptTypes, ServiceStatus, FlexOptionsNames, JobStatus } = require('../Constants');
 const CommandResult = require('../CommandResult');
 const KinveyError = require('../KinveyError');
-const { getCommandNameFromOptions, isNullOrUndefined } = require('../Utils');
+const { formatList, getCommandNameFromOptions, isEmpty, isValidNonZeroInteger, isValidTimestamp, isNullOrUndefined } = require('../Utils');
 
 /**
  * Handles flex-related logic.
@@ -31,13 +33,15 @@ class FlexController extends BaseController {
    * Creates an instance.
    * @param {Object} options
    * @param {Object} options.cliManager
-   * @param {Object} options.flexService
+   * @param {Object} options.applicationsService
+   * @param {Object} options.organizationsService
+   * @param {Object} options.servicesService
    */
   constructor(options) {
     super(options);
-    this.flexService = options.flexService;
     this.applicationsService = options.applicationsService;
     this.organizationsService = options.organizationsService;
+    this.servicesService = options.servicesService;
     this.projectSetup = new ProjectSetup(this.cliManager.config.paths.project);
     this._metadata = {
       serviceName: null,
@@ -92,6 +96,24 @@ class FlexController extends BaseController {
     const typeMsg = isPositionalArg ? 'positional arguments' : 'options';
     const errMsg = `${Errors.ProjectNotConfigured.MESSAGE} Alternatively, use ${typeMsg}: ${keys.join(', ')}.`;
     return new KinveyError(Errors.ProjectNotConfigured.NAME, errMsg);
+  }
+
+  static paintServiceStatus(status) {
+    let paintedStatus;
+
+    if (status === ServiceStatus.ONLINE) {
+      paintedStatus = chalk.greenBright(ServiceStatus.ONLINE);
+    } else if (status === ServiceStatus.UPDATING) {
+      paintedStatus = chalk.yellowBright(ServiceStatus.UPDATING);
+    } else if (status === ServiceStatus.NEW) {
+      paintedStatus = chalk.cyanBright(ServiceStatus.NEW);
+    } else if (status === ServiceStatus.ERROR) {
+      paintedStatus = chalk.redBright(ServiceStatus.ERROR);
+    } else {
+      paintedStatus = status; // shouldn't happen
+    }
+
+    return paintedStatus;
   }
 
   /**
@@ -151,6 +173,191 @@ class FlexController extends BaseController {
   }
 
   /**
+   * Gets all apps and all orgs. If there are no apps and no orgs at the same time, passes error to callback. If entities exist, passes an object with 'app' and 'org' properties.
+   * @param done
+   * @private
+   */
+  _getAllDomainEntities(done) {
+    const result = {};
+
+    async.series([
+      (next) => {
+        this.organizationsService.getAll((err, orgs) => {
+          if (err) {
+            return next(err);
+          }
+
+          if (!isEmpty(orgs)) {
+            result[DomainTypes.ORG] = orgs;
+          }
+
+          next();
+        });
+      },
+      (next) => {
+        this.applicationsService.getAll((err, apps) => {
+          if (err) {
+            return next(err);
+          }
+
+          if (!isEmpty(apps)) {
+            result[DomainTypes.APP] = apps;
+          }
+
+          next();
+        });
+      }
+    ], (err) => {
+      if (err) {
+        return done(err);
+      }
+
+      if (isEmpty(result)) {
+        return done(new KinveyError(Errors.NoAppsAndOrgsFound));
+      }
+
+      done(null, result);
+    });
+  }
+
+  /**
+   * Gets the user's choice of domain: app or org.
+   * @param done
+   * @private
+   */
+  _getDomainTypeChoice(done) {
+    const domainOpts = [{ name: 'App' }, { name: 'Organization' }];
+    const domainQuestion = this.cliManager.prompter.buildQuestion(PromptMessages.INPUT_DOMAIN, 'domain', PromptTypes.LIST, null, domainOpts);
+
+    this.cliManager.log(LogLevel.DEBUG, InfoMessages.APP_OR_ORG_PROMPTING);
+    this.cliManager.prompter.prompt(domainQuestion, (err, data) => {
+      if (err) {
+        return done(err);
+      }
+
+      let result;
+      if (data.domain === 'App') {
+        result = DomainTypes.APP;
+      } else {
+        result = DomainTypes.ORG;
+      }
+
+      done(null, result);
+    });
+  }
+
+  /**
+   * Gets the user's choice of specific domain entity (specific app or org).
+   * @param {Array} entities
+   * @param {Constants.DomainTypes} domainType
+   * @param done
+   * @private
+   */
+  _getDomainEntityChoice(entities, domainType, done) {
+    const isAppDomain = domainType === DomainTypes.APP;
+    const logMsg = isAppDomain ? InfoMessages.APP_PROMPTING : InfoMessages.ORG_PROMPTING;
+    this.cliManager.log(LogLevel.DEBUG, logMsg);
+    const msg = isAppDomain ? PromptMessages.INPUT_APP : PromptMessages.INPUT_ORG;
+    const question = this.cliManager.prompter.buildQuestion(msg, 'entity', PromptTypes.LIST, null, formatList(entities));
+    this.cliManager.prompter.prompt(question, (err, data) => {
+      if (err) {
+        return done(err);
+      }
+
+      done(null, data.entity);
+    });
+  }
+
+  /**
+   * Gets the user's choice of service.
+   * @param {Array|Object} services
+   * @param done
+   * @private
+   */
+  _getServiceChoice(services, done) {
+    if (!Array.isArray(services)) {
+      services = [services];
+    }
+
+    this.cliManager.log(LogLevel.DEBUG, InfoMessages.SERVICE_PROMPTING);
+    const question = this.cliManager.prompter.buildQuestion(PromptMessages.INPUT_SPECIFIC_SERVICE, 'service', PromptTypes.LIST, null, formatList(services));
+    this.cliManager.prompter.prompt(question, (err, data) => {
+      if (err) {
+        return done(err);
+      }
+
+      done(null, data.service);
+    });
+  }
+
+  /**
+   * Gathers the required data to setup a flex service.
+   * @param done
+   * @private
+   */
+  _getInitInput(done) {
+    const input = {};
+    let allDomainEntities;
+
+    async.waterfall([
+      (next) => {
+        this._getAllDomainEntities((err, data) => {
+          if (err) {
+            return next(err);
+          }
+
+          allDomainEntities = data;
+          next();
+        });
+      },
+      (next) => {
+        const canChoose = Object.keys(allDomainEntities).length > 1;
+        if (!canChoose) {
+          const domainType = Object.keys(allDomainEntities)[0];
+          input.domain = domainType;
+          return setImmediate(() => { next(null, domainType); });
+        }
+
+        this._getDomainTypeChoice((err, domain) => {
+          if (err) {
+            return next(err);
+          }
+
+          input.domain = domain;
+          next(null, domain);
+        });
+      },
+      (domain, next) => {
+        const entities = allDomainEntities[domain];
+        this._getDomainEntityChoice(entities, domain, next);
+      },
+      (entity, next) => {
+        entity.schemaVersion = entity.schemaVersion || this.cliManager.config.defaultSchemaVersion;
+        input.schemaVersion = entity.schemaVersion;
+        input.domainEntityId = entity.id;
+        this.servicesService.getAllInternalFlexServicesByDomain(input.domain, entity, next);
+      },
+      (services, next) => {
+        this._getServiceChoice(services, (err, data) => {
+          if (err) {
+            return next(err);
+          }
+
+          input.serviceId = data.id;
+          input.serviceName = data.name;
+          next(null);
+        });
+      }
+    ], (err) => {
+      if (err) {
+        return done(err);
+      }
+
+      done(null, input);
+    });
+  }
+
+  /**
    * Handles the 'flex init' command.
    * @param {Object} options
    * @param done
@@ -166,7 +373,7 @@ class FlexController extends BaseController {
         setImmediate(() => next(err));
       },
       (next) => {
-        this.flexService.getInitInput((err, input) => {
+        this._getInitInput((err, input) => {
           if (err) {
             return next(err);
           }
@@ -191,25 +398,13 @@ class FlexController extends BaseController {
 
   deploy(options, done) {
     const dir = this.cliManager.config.paths.package;
-    let version;
     let jobId;
+
     async.series([
       (next) => {
-        this.flexService.validateProject(dir, (err, pkgVersion) => {
-          if (err) {
-            return next(err);
-          }
-
-          version = pkgVersion;
-          next();
-        });
-      },
-      (next) => {
-        const serviceId = this._metadata.serviceId;
         const schema = this._metadata.schemaVersion;
-        const artifacts = this.cliManager.config.artifacts;
-        const uploadTimeout = this.cliManager.config.flexProjectUploadTimeout;
-        this.flexService.deployProject(dir, serviceId, version, schema, uploadTimeout, FlexProjectMaxSize, artifacts, (err, deployJobId) => {
+        const serviceId = this._metadata.serviceId;
+        this.servicesService.deployFlexProject(dir, serviceId, schema, (err, deployJobId) => {
           if (err) {
             return next(err);
           }
@@ -242,7 +437,7 @@ class FlexController extends BaseController {
   }
 
   job(options, done) {
-    this.flexService.jobStatus(this._metadata.jobId, this._metadata.schemaVersion, (err, result) => {
+    this.servicesService.jobStatus(this._metadata.jobId, this._metadata.schemaVersion, (err, result) => {
       if (err) {
         return done(err);
       }
@@ -260,13 +455,38 @@ class FlexController extends BaseController {
     });
   }
 
+  // eslint-disable-next-line class-methods-use-this
+  getTableDataFromStatusData(data, { serviceId, serviceName }) {
+    const result = {};
+
+    result.status = FlexController.paintServiceStatus(data.status);
+    result.version = data.version;
+    result.id = serviceId;
+
+    if (!isNullOrUndefined(serviceName)) {
+      result.name = serviceName;
+    }
+
+    const hasDeployInfo = data.requestedAt && !isEmpty(data.deployUserInfo);
+    if (hasDeployInfo) {
+      result.requestedAt = moment(data.requestedAt).format('dddd, MMMM Do YYYY, h:mm:ss A');
+      result.deployerEmail = data.deployUserInfo.email;
+
+      if (data.deployUserInfo.firstName && data.deployUserInfo.lastName) {
+        result.deployerName = `${data.deployUserInfo.firstName} ${data.deployUserInfo.lastName}`;
+      }
+    }
+
+    return result;
+  }
+
   status(options, done) {
-    this.flexService.getServiceStatus(this._metadata, (err, result) => {
+    this.servicesService.getServiceStatus(this._metadata, (err, result) => {
       if (err) {
         return done(err);
       }
 
-      const tableData = this.flexService.getTableDataFromStatusData(result, this._metadata);
+      const tableData = this.getTableDataFromStatusData(result, this._metadata);
       const cmdResult = new CommandResult()
         .setRawData(result)
         .setTableData(tableData);
@@ -299,7 +519,7 @@ class FlexController extends BaseController {
           backingServers: [{ secret: options[FlexOptionsNames.SERVICE_SECRET] }]
         };
 
-        this.flexService.createService(data, entity.id, domainType, next);
+        this.servicesService.create(data, entity.id, domainType, next);
       }
     ], (err, result) => {
       if (err) {
@@ -319,7 +539,7 @@ class FlexController extends BaseController {
       schemaVersion: this._metadata.schemaVersion
     };
 
-    this.flexService.getServices(this._metadata.domain, domainEntity, (err, services) => {
+    this.servicesService.getAllInternalFlexServicesByDomain(this._metadata.domain, domainEntity, (err, services) => {
       if (err) {
         return done(err);
       }
@@ -336,13 +556,62 @@ class FlexController extends BaseController {
     });
   }
 
+  // eslint-disable-next-line class-methods-use-this
+  validateLogsOptions({ from, to, number, page }, done) {
+    let err = null;
+
+    const fromExists = !isNullOrUndefined(from);
+    const toExists = !isNullOrUndefined(to);
+
+    if (fromExists && !isValidTimestamp(from)) {
+      err = new KinveyError('InvalidParameter', `Logs 'from' flag ${LogErrorMessages.INVALID_TIMESTAMP}`);
+    } else if (toExists && !isValidTimestamp(to)) {
+      err = new KinveyError('InvalidParameter', `Logs 'to' flag ${LogErrorMessages.INVALID_TIMESTAMP}`);
+    } else if (!isNullOrUndefined(number) && !isValidNonZeroInteger(number)) {
+      err = new KinveyError('InvalidParameter', `Logs 'number' flag ${LogErrorMessages.INVALID_NONZEROINT}`);
+    } else if (!isNullOrUndefined(page) && !isValidNonZeroInteger(page)) {
+      err = new KinveyError('InvalidParameter', `Logs 'page' flag ${LogErrorMessages.INVALID_NONZEROINT}`);
+    } else if ((fromExists && toExists) && !moment(from).isBefore(to)) {
+      err = new KinveyError('InvalidParameter', `'${FlexOptionsNames.FROM}' timestamp must be before '${FlexOptionsNames.TO}' timestamp.`);
+    }
+
+    done(err);
+  }
+
+  // eslint-disable-next-line class-methods-use-this
+  getTableDataFromLogs(logs) {
+    const result = [];
+
+    logs.forEach((x) => {
+      if (isEmpty(x)) {
+        return;
+      }
+
+      const log = {
+        containerId: x.containerId.substring(0, 12)
+      };
+
+      if (!isNullOrUndefined(x.message)) {
+        // remove EOL characters
+        log.message = (x.message.replace(/\r?\n|\r/g, ' ')).trim();
+      }
+
+      log.threshold = x.threshold;
+      log.timestamp = x.timestamp;
+
+      result.push(log);
+    });
+
+    return result;
+  }
+
   logs(options, done) {
     async.series([
       (next) => {
-        this.flexService.validateLogsOptions(options, next);
+        this.validateLogsOptions(options, next);
       },
       (next) => {
-        this.flexService.getServiceLogs(options, this._metadata.serviceId, this._metadata.schemaVersion, next);
+        this.servicesService.getServiceLogs(options, this._metadata.serviceId, this._metadata.schemaVersion, next);
       },
     ], (err, results) => {
       if (err) {
@@ -350,7 +619,7 @@ class FlexController extends BaseController {
       }
 
       const result = results.pop();
-      const tableData = this.flexService.getTableDataFromLogs(result);
+      const tableData = this.getTableDataFromLogs(result);
       const cmdResult = new CommandResult()
         .setRawData(result)
         .setTableData(tableData);
@@ -362,7 +631,7 @@ class FlexController extends BaseController {
     let jobId;
     async.series([
       (next) => {
-        this.flexService.execRecycle(this._metadata.schemaVersion, this._metadata.serviceId, (err, recycleJobId) => {
+        this.servicesService.execRecycle(this._metadata.schemaVersion, this._metadata.serviceId, (err, recycleJobId) => {
           if (err) {
             return next(err);
           }
@@ -401,7 +670,7 @@ class FlexController extends BaseController {
         super.confirmDeleteOperation(options[CommonOptionsNames.NO_PROMPT], EntityType.SERVICE, serviceId, next);
       },
       (next) => {
-        this.flexService.deleteServiceById(serviceId, next);
+        this.servicesService.deleteInternalFlexServiceById(serviceId, next);
       },
       (next) => {
         const currentProfile = this.cliManager.getCurrentProfileName();
