+++ lib/service/ServicesController.js
@@ -17,9 +17,9 @@ const async = require('async');
 
 const BaseController = require('../BaseController');
 const CommandResult = require('../CommandResult');
-const { ActiveItemType, AppOptionsName, CommonOptionsNames, ConfigType, DomainTypes, EntityType, EnvOptionsName, Errors, Mapping, OperationType, OrgOptionsName } = require('../Constants');
+const { AppOptionsName, ConfigType, DomainTypes, EntityType, OperationType, OrgOptionsName } = require('../Constants');
 const KinveyError = require('../KinveyError');
-const { getCommandNameFromOptions, isNullOrUndefined, mapFromSource, sortList } = require('../Utils');
+const { isNullOrUndefined } = require('../Utils');
 
 class ServicesController extends BaseController {
   constructor(options) {
@@ -28,13 +28,25 @@ class ServicesController extends BaseController {
     this.organizationsService = options.organizationsService;
   }
 
+  static getCommandResult(operationType, service) {
+    const rawData = { id: service.id };
+    const cmdResult = new CommandResult().setBasicMsg(operationType, EntityType.SERVICE, service.id);
+
+    if (!isNullOrUndefined(service.jobId)) {
+      rawData.jobId = service.jobId;
+      cmdResult.setAdditionalMsg(`Job ID: ${service.jobId}`);
+    }
+
+    cmdResult.setRawData(rawData);
+    return cmdResult;
+  }
+
   create(options, done) {
     let domainType;
     let domainId;
 
     async.series([
       (next) => {
-        // TODO:
         const appIsSet = !isNullOrUndefined(options[AppOptionsName.APP]);
         const orgIsSet = !isNullOrUndefined(options[OrgOptionsName.ORG]);
         if (appIsSet) {
@@ -49,7 +61,7 @@ class ServicesController extends BaseController {
           });
         } else if (orgIsSet) {
           domainType = DomainTypes.ORG;
-          this.organizationsService.getByIdOrName(options[OrgOptionsName.ORG],  (err, data) => {
+          this.organizationsService.getByIdOrName(options[OrgOptionsName.ORG], (err, data) => {
             if (err) {
               return next(err);
             }
@@ -79,9 +91,7 @@ class ServicesController extends BaseController {
       }
 
       const result = results.pop();
-      const cmdResult = new CommandResult()
-        .setRawData({ id: result.id })
-        .setBasicMsg(OperationType.CREATE, EntityType.SERVICE, result.id);
+      const cmdResult = ServicesController.getCommandResult(OperationType.CREATE, result);
       done(null, cmdResult);
     });
   }
@@ -99,9 +109,7 @@ class ServicesController extends BaseController {
         return done(err);
       }
 
-      const cmdResult = new CommandResult()
-        .setRawData({ id: data.id })
-        .setBasicMsg(OperationType.UPDATE, EntityType.SERVICE);
+      const cmdResult = ServicesController.getCommandResult(OperationType.UPDATE, data);
       done(null, cmdResult);
     });
   }
