+++ lib/service/servicesRouter.js
@@ -13,7 +13,7 @@
  * contents is a violation of applicable laws.
  */
 
-const { AppOptions, AppOptionsName, CommandRequirement, Namespace, OrgOptions, OrgOptionsName } = require('./../Constants');
+const { AppOptions, AppOptionsName, CommandRequirement, FlexOptionsNames, FlexOptions, Namespace, OrgOptions, OrgOptionsName } = require('./../Constants');
 
 const cmdDefinitions = [
   {
@@ -31,11 +31,11 @@ const cmdDefinitions = [
     requirements: [CommandRequirement.AUTH]
   },
   {
-    command: `push <serviceId> <file>`,
+    command: `push <${FlexOptionsNames.SERVICE_ID}> <file>`,
     desc: 'Update a service',
     builder: (commandsManager) => {
       commandsManager
-        .positional('serviceId', { describe: 'Service ID' })
+        .positional(FlexOptionsNames.SERVICE_ID, FlexOptions[FlexOptionsNames.SERVICE_ID])
         .positional('file', { describe: 'Path to a service configuration file' });
     },
     handlerName: 'update',
