+++ test/integration/commands/flex/flex-deploy.test.js
@@ -20,7 +20,7 @@ const path = require('path');
 
 const { AuthOptionsNames } = require('./../../../../lib/Constants');
 const FlexController = require('./../../../../lib/flex/FlexController');
-const FlexService = require('./../../../../lib/flex/FlexService');
+const ServicesService = require('./../../../../lib/service/ServicesService');
 const CLIManager = require('./../../../../lib/CLIManager');
 const logger = require('./../../../../lib/logger');
 const Setup = require('./../../../../lib/Setup');
@@ -127,6 +127,7 @@ describe(`${baseCmd}`, () => {
           const manager = new CLIManager({ setup, config: testsConfig, logger, commandsManager: yargs });
           manager.sendRequest = function stubSendRequest(options, done) {
             const mockReq = {
+              on: () => {},
               once: () => {},
               send: () => {
                 if (options.endpoint === Endpoints.session()) {
@@ -158,8 +159,8 @@ describe(`${baseCmd}`, () => {
             return mockReq;
           };
 
-          const flexService = new FlexService(manager);
-          const ctrl = new FlexController({ cliManager: manager, flexService });
+          const servicesService = new ServicesService(manager);
+          const ctrl = new FlexController({ cliManager: manager, servicesService });
 
           const options = {
             [AuthOptionsNames.EMAIL]: existentUserOne.email,
